==> 3. Download information .ipynb <==
## Descarga de los archivos

Librerías y paths utilizados

#%pip install -U BeautifulSoup4
#Para web scraping
from bs4 import BeautifulSoup as bs
import requests

#Para descargar los archivos
import pandas as pd
import numpy as np

#De donde se bajaran los archivos
URL_s = "https://raw.githubusercontent.com/srovere/nearbpo-data-task-level1/main/data/stations.csv"
URL_v = "https://raw.githubusercontent.com/srovere/nearbpo-data-task-level1/main/data/variables.csv"
URL_o = ['http://www.inia.org.uy/online/site/gras_datos.php?filtro=1&fecha_des=2019-01-01&fecha_has=2021-12-30&ver=20000&campos=15',
         'http://www.inia.org.uy/online/site/gras_datos.php?filtro=2&fecha_des=2019-01-01&fecha_has=2021-12-30&ver=20000&campos=15',
         'http://www.inia.org.uy/online/site/gras_datos.php?filtro=3&fecha_des=2019-01-01&fecha_has=2021-12-30&ver=20000&campos=15',
         'http://www.inia.org.uy/online/site/gras_datos.php?filtro=4&fecha_des=2019-01-01&fecha_has=2021-12-30&ver=20000&campos=15',
         'http://www.inia.org.uy/online/site/gras_datos.php?filtro=5&fecha_des=2019-01-01&fecha_has=2021-12-30&ver=20000&campos=15']

#Donde se guardaran los archivos
        #cambiar al directorio/ruta deseado.
path = "C:\Users\ppyan\.jupyter\"



#### Descargando el primer archivo

descarga1 = pd.read_csv(URL_v) #Se obtiene la url
#display(descarga1)           #Revisar los datos
descarga1.to_csv("variables.csv", encoding='utf-8',index= False) #Se descarga como csv
print("Se descargo el csv")

#### Descargando el segundo archivo

descarga2 = pd.read_csv(URL_s) #Se obtiene la url
#display(descarga2)           #Revisar los datos
descarga2.to_csv("stations.csv", encoding='utf-8',index= False) #Se descarga como csv
print("Se descargo el csv")

### Descargando los archivos de las 5 estaciones

#### ejemplo de las descargas station **NO EJECUTAR IRSE A LAS DESCARGAS**

conexión a la url

captura = requests.get(URL_o)  #se obtiene la url (manda solicitud en tipo http)
conecta = captura.content      #entrando al codigo fuente de la pagina
conexion = bs(conecta,'html.parser') #listo para buscar y leer formato html
#print(conexion) #vista previa de la web

#Los datos que se requieren estan en: </tr> y <td

Visualización de los datos

#for i in conexion.find_all("tr"): #muestra en la sección en donde estan los datos
limite = 0  #inicialización de contador
Encontrar = conexion.find_all("td")   #encuentra las secciones con 'td'
         #una ves en la sección ir guardando SOLO el texto (los datos) hasta que el contador llegue a 5
x = [j.text for j in Encontrar if limite < 5] 
                #Los datos recuperados se pasan a un dataframe 
df = pd.DataFrame(x)
df

limpieza  y descarga de los datos

#df.tail()          #Se localiza en donde estan los datos requeridos (hasta el index 2202)

#se localiza exactamente las temperaturas y grados
datos = df.iloc[13:2203] #2203-(1095x2) = 13

datos_f = datos.iloc[::2] #Las fechas estan en filas pares, se separan
datos_t = datos.iloc[1::2] #Las temperaturas estan en filas impares, se separan


df_f = pd.DataFrame(datos_f) #se vuelve dataframe a las fechas
df_f.set_index(np.arange(1,1096),inplace=True) #se reinicia el index
df_f.index.name = 'filas' #se coloca nombre a la columna

df_t = pd.DataFrame(datos_t) #se vuelve dataframe a las temperaturas
df_t.set_index(np.arange(1,1096),inplace=True) #se reinicia el index
df_t.index.name = 'filas' #se coloca nombre a la columna


clima = pd.merge(df_f,df_t, on=['filas']) #se unen los 2 df por medio del index
clima.columns = ['Fechas','Temperatura Máxima ºC'] #se colocan los headers, respectivamente
#clima['Fechas'].astype('datetime64') #se colocan las fechas como tipo de dato datetime
clima['Temperatura Máxima ºC'].astype('float64') #se colocan las temperaturas como tipo de dato float

#Descarga
clima.to_csv('observations.csv',encoding='utf-8',index=False) #se descarga el df como csv, con codigicacion utf8 y sin index
print("Archivo descargado")

#### **DESCARGA** de stations 1, 2 , 3, 4 y 5

def des(URL):  #se ejecuta 5 veces
  captura2 = requests.get(URL)  #se obtiene la url (manda solicitud en tipo http)
  conecta2 = captura2.content      #entrando al codigo fuente de la pagina
  conexion2 = bs(conecta2,'html.parser') #listo para buscar y leer formato html

  limite2 = 0  #inicialización de contador
  Encontrar2 = conexion2.find_all("td")   #encuentra las secciones con 'td'
          #una ves en la sección ir guardando SOLO el texto (los datos) hasta que el contador llegue a 5
  x2 = [j.text for j in Encontrar2 if limite2 < 5] 
                  #Los datos recuperados se pasan a un dataframe 
  df2 = pd.DataFrame(x2)

  datos2 = df2.iloc[13:2203] #2203-(1095x2) = 13

  datos_f2 = datos2.iloc[::2] #Las fechas estan en filas pares, se separan
  datos_t2 = datos2.iloc[1::2] #Las temperaturas estan en filas impares, se separan


  df_f2 = pd.DataFrame(datos_f2) #se vuelve dataframe a las fechas
  df_f2.set_index(np.arange(1,1096),inplace=True) #se reinicia el index
  df_f2.index.name = 'filas' #se coloca nombre a la columna

  df_t2 = pd.DataFrame(datos_t2) #se vuelve dataframe a las temperaturas
  df_t2.set_index(np.arange(1,1096),inplace=True) #se reinicia el index
  df_t2.index.name = 'filas' #se coloca nombre a la columna


  clima2 = pd.merge(df_f2,df_t2, on=['filas']) #se unen los 2 df por medio del index
  clima2.columns = ['Fechas','Temperatura Máxima ºC'] #se colocan los headers, respectivamente
  #clima2['Fechas'].astype('datetime64') #se colocan las fechas como tipo de dato datetime
  #clima2['Temperatura Máxima ºC'].astype('float64') #se colocan las temperaturas como tipo de dato float ->comentar para las station2,3,4, ya que les falta almenos 1 temperatura
#Descarga
  clima2.to_csv('observations'+URL[57:58]+'.csv',encoding='utf-8',index=False) #se descarga el df como csv, con codigicacion utf8 y sin index
  print("Archivo"+URL[57:58]+" descargado")

def desc(LINK):
  for i in LINK: #para cada una de las 5 url's
    des(i)       #ejecuta web scrapping
desc(URL_o)     #se le pasa la lista de las URL's
