==> 4. Import data.ipynb <==
## Carga de los archivos a la base de datos

Librerías y paths utilizados

#%pip install pyodbc
import pandas as pd
import numpy as np

#Donde se guardaran los archivos
        #cambiar al directorio/ruta deseado.
path = "C:\Users\ppyan\.jupyter\"

### Carga en memoria

variables

df_v = pd.read_csv('variables.csv')
df_v

stations

df_s = pd.read_csv('stations.csv')
df_s

datos de las 5 stations

df_o1 = pd.read_csv('observations1.csv',encoding= 'utf-8')
df_o2 = pd.read_csv('observations2.csv',encoding= 'utf-8')
df_o3 = pd.read_csv('observations3.csv',encoding= 'utf-8')
df_o4 = pd.read_csv('observations4.csv',encoding= 'utf-8')
df_o5 = pd.read_csv('observations5.csv',encoding= 'utf-8')

reemplazo de valores faltantes de algunas tablas

#df_o2['Temperatura Máxima ºC'].isna().value_counts()
df_o2.fillna(0,inplace=True)
df_o3.fillna(0,inplace=True)
df_o4.fillna(0,inplace=True)

### Conexión a la base de datos y subida de archivos

importando variables

import pyodbc


connection = pyodbc.connect('DRIVER={SQL Server};SERVER=LAPTOP-FAKK9GHA;DATABASE=weather;UID=NearBPO;PWD=password')
cursor = connection.cursor()
#Inserting data in SQL Table:- 
for index,row in df_v.iterrows():
    cursor.execute("INSERT INTO variables(variable_id,name,description, units) values (?,?,?,?)", row['variable_id'],row['name'],row['description'],row['units']) 
connection.commit()
cursor.close()
#connection.close()

importando stations

connection2 = pyodbc.connect('DRIVER={SQL Server};SERVER=LAPTOP-FAKK9GHA;DATABASE=weather;UID=NearBPO;PWD=password')
cursor2 = connection2.cursor()
for index,row in df_s.iterrows():
    cursor2.execute("INSERT INTO stations(station_id,name,latitude, longitude, elevation) values (?,?,?,?,?)", row['station_id'],row['name'],row['latitude'],row['longitude'], row['elevation']) 
connection2.commit()
cursor2.close()
#connection.close()

importando a observations

connection3 = pyodbc.connect('DRIVER={SQL Server};SERVER=LAPTOP-FAKK9GHA;DATABASE=weather;UID=NearBPO;PWD=password')
cursor3 = connection3.cursor()
for index,row in df_o5.iterrows(): #Ir cambiando de df_o hasta df_o5
    #cursor3.execute("TRUNCATE TABLE observations")
    cursor3.execute("INSERT INTO observations(observed_value) values (?)", row['Temperatura Máxima ºC'])
    #cursor3.execute("INSERT INTO observations (sensor_id) SELECT station_id FROM stations,observations WHERE stations.station_id =1 AND  observations.observed_value>1")
connection3.commit()
cursor3.close()
