==> 5. Menú.ipynb <==
# CLI

import pyodbc  #Para la conexión con sql server

#función para crear
def c():
    print("formato esperado: CREATE TABLE table_name (column1 datatype,column2 datatype,column3 datatype,....);")
    tabla = str(input("ingresar query para crear tabla: "))
    connection = pyodbc.connect('DRIVER={SQL Server};SERVER=LAPTOP-FAKK9GHA;DATABASE=weather;UID=NearBPO;PWD=password')
    cursor = connection.cursor() 
    try:
        cursor.execute(tabla)

        connection.commit()
        cursor.close()
    except (NameError,pyodbc.ProgrammingError):
        print("query invalido, intentalo de nuevo.\n")
        return c()
    
    print("Verificar los cambios en la base de datos.")
    return opcion

#función para select
def r():
    print("formato esperado: SELECT column1, column2, ...FROM table_name;")
    select = str(input("ingresar query: "))
    connection = pyodbc.connect('DRIVER={SQL Server};SERVER=LAPTOP-FAKK9GHA;DATABASE=weather;UID=NearBPO;PWD=password')
    cursor = connection.cursor() 
    try:
        cursor.execute(select)
        
        connection.commit()
        cursor.close()
    except (NameError,pyodbc.ProgrammingError):
        print("query invalido, intentalo de nuevo.\n")
        return r()
    
    print("Verificar los cambios en la base de datos.")
    return opcion

#función para insertar
def u():
    print("""formatos esperados: \t INSERT INTO table_name (column1, column2, column3, ...)VALUES (value1, value2, value3, ...);
                                 UPDATE table_name SET column1 = value1, column2 = value2, ...WHERE condition;""")
    update = str(input("Ingresar query para insertar elementos: "))
    connection = pyodbc.connect('DRIVER={SQL Server};SERVER=LAPTOP-FAKK9GHA;DATABASE=weather;UID=NearBPO;PWD=password')
    try:
        cursor = connection.cursor() 
        cursor.execute(update)

        connection.commit()
        cursor.close()
    except (NameError,pyodbc.ProgrammingError):
        print("query invalido, intentalo de nuevo.\n")
        return u()
    
    print("Verificar los cambios en la base de datos.")
    return opcion

#función para eliminar
def d():
    print("""formatos esperados: \t\t DELETE FROM table_name WHERE condition;
                                 DROP TABLE table_name;
                                 DROP DATABASE databasename;
                                 TRUNCATE TABLE table_name;
                                 """)
    delete = str(input("Ingresar query para eliminar elemento(s): "))
    connection = pyodbc.connect('DRIVER={SQL Server};SERVER=LAPTOP-FAKK9GHA;DATABASE=weather;UID=NearBPO;PWD=password')
    try:
        cursor = connection.cursor() 
        cursor.execute(delete)

        connection.commit()
        cursor.close()
    except (NameError,pyodbc.ProgrammingError):
        print("\nquery invalido, intentalo de nuevo.\n")
        return d()
    
    print("Verificar los cambios en la base de datos.")
    return opcion



opcion = '0' #variable opción

#Muestra el menú
print("""\n\n\t\t\t Menú de la base de datos weather. 
         \t\t Elige una de las siguientes opciones\n""")

print("""\tC. Create table(s) \t  press 1)
        R. Read/select            press 2)
        U. Update/insert table(s) press 3)
        D. Delete/Drop/Truncate   press 4)
        1. Exit                   press 5)""")

#ciclo para el menú
while not(opcion=='5'): #mientras no se salga..

    opcion=input('\n  --- ¿Cuál opcion?: ') 
    
    if (opcion=='1'):
        print(' **** ha elegido Create ****')
        c()
    elif (opcion=='2'):
        print(' **** ha elegido Read ****')
        r()
    elif (opcion=='3'):
        print(' **** ha elegido Update ****')
        u()
    elif (opcion=='4'):
        print(' **** ha elegido Delete ****')
        d()   
    elif (opcion=='5'):
        print('\n **** Saliendo del menú.....  ****')
        print(' **** Vuelva pronto ****\n')
    else:
        print('\t\nNo existe esa opcion')
